requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.17.0
Flask==3.1.0
fsspec==2025.3.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.7
openai==1.65.4
openai-whisper==20240930
playsound==1.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
python==3.10.9